More micro commits for Event restructure
I have been commiting small amounts of changes to the event structure of OpMon. This is by no means going to build or anything. I'm trying to get an idea of how this will fit into the bigger picture.

I am trying to get a sense of what a clean interface could look like then I have to somehow figure out how to wire it back into the rest of the architecture of the game. I think as long as I dont change the controllers too
much I should be able to come back and accomplish this...