Added new Test events to maps.json and changed Sound event ctor params
I ran into a kind of big problem right before commiting. The DoorEvent is a new macro event which is just a composition of regular events. However, when trying to add this new
macro type event to the map on the map.cpp file.. I can no longer insert it into the event container that the map has. I'm not sure how to solve this one yet...